fix(scoring/admin): 分数 float 传递、收卷置会话状态与遗留项清理
评分正确性:
- 客观分去掉 int() 截断 (computeObjective / gradeObjectiveInTx /
  CreateFromSessionTx 全链路 float64): 题库 14 张卷各有 20 道 1.5 分单选,
  截断会永久丢 0.5 分; DB 列本为 double precision, worker 与复核已是浮点
- 接线 scoring.multiple_choice_partial: 注入 submissions.Service (HTTP
  主路径) 与 Repository (finalize 事务内判分), 不再硬编码 partial=false;
  契约 harness 配置同步开启与生产对齐

收卷闭环:
- FinalizeRun 同事务把自动收卷的 session 批量置 submitted:
  否则学生端轮询永远等不到 submission_id, 收卷后卡"正在自动收卷"页

管理端:
- 提交列表 LEFT JOIN exam_runs 补前端渲染的 7 字段 (paper_name/round_no/
  department/objective_score/subjective_score_final/auto_submit_reason/
  run_is_legacy), 实现 keyword (姓名/工号模糊) 与 paper_id 过滤
- stats low_confidence_count 改按 grading_detail 单题低置信标志计数
  (worker 整卷状态无 low_confidence 值, 原统计恒 0)
- 前端补 partial_manual (部分人工) 状态徽章

配置:
- config.dev.yaml 键名对齐 Go 结构体 (max_conns/min_conns/
  connect_timeout_seconds/statement_timeout_seconds), 删除 yaml.v3
  静默忽略的死段, 注明 5432/5433 端口陷阱

实测: 1.5+2.0(多选半对)=3.5 不截断; 收卷 3s 内 session=submitted +
admin_closed 草稿判 1.5; 列表字段/过滤、低置信计数、dev 配置解析均验证.
回归: go test 14 包全过; 契约 25 passed+8 xfailed; pytest 147 passed.

+++ internal/submissions/types.go
@@ -75,8 +75,9 @@ type PreparedSubmission struct {
 	AutoSubmitReason  *string
 
 	// 客观题即时判分结果 (事务外算好)。主观题请求占位为 0, 待 worker 填充。
-	ObjectiveScore     int
-	ObjectiveMaxScore   int
+	// float64: 题库存在 1.5 分制单选, int 截断会永久丢 0.5 分 (DB 列 double precision).
+	ObjectiveScore     float64
+	ObjectiveMaxScore   float64
 	ObjectiveDetailJSON []byte // 单题判分明细 (array of objective.GradeResult 字段)
 	QuestionTotal       int
 }
@@ -106,7 +107,7 @@ type Submission struct {
 	ID                string
 	GradingStatus     Status
 	ReviewStatus      ReviewStatus
-	ObjectiveScore    int
+	ObjectiveScore    float64
 }
 
 // ErrSubmissionClosed 表示 session / run 已关闭，不允许提交。
